Added sources from irc and changed a bit on libs, read on!
irc sources where added to the serveur sources to have a base to start on, I believe adding type such as FD_UNKNOWN would be great in order to differentiate from FD_FREE.
I modified libftsock so that it accept the port 0 for the server and get back the port assigned in the structure t_tcpsock. The irc serveur display the port used if it was passed 0 as a parameter.
I also added to the libft the following:
- ft_free_array, used to free the splited command by the serveur.
- ft_isx, a facility that take a string and a ft_is function, and check if the sting is of the type checked by the function passed in parameter.
last but not least, the makefile wasn't changed, only serveur.mk. It compile and link successfully.
gg wall-o-text