added client_unknown, a function that check the first response from the clients and determine what function will be used in the future.
We need to implement client_player and client_graphic, facilities that will interact with their respective type of client. client_write is generic, it is only called when data need to be sent, no need to update it, except to verify that it properly check for the return value of send() and move pointers to the buffer accordingly.

Catching of signal such as ^C should also considered